Made better use of expression sequence
